refactor: centralize prompt utilities
